+++ Raytracer/Equations.cs
@@ -8,30 +8,30 @@ namespace Raytracer
 {
     struct LinearEquation
     {
-        public float slope;
-        public float intercept;
+        public double slope;
+        public double intercept;
 
-        public LinearEquation(float slope, float intercept)
+        public LinearEquation(double slope, double intercept)
         {
             this.slope = slope;
             this.intercept = intercept;
         }
 
-        public float ValueAt(float t)
+        public double ValueAt(double t)
         {
             return (t * slope) + intercept;
         }
 
-        public float SolveWhenValueIs(float value)
+        public double SolveWhenValueIs(double value)
         {
             return (value - intercept) / slope;
         }
 
-        public static LinearEquation operator +(LinearEquation eqn, float value)
+        public static LinearEquation operator +(LinearEquation eqn, double value)
         {
             return new LinearEquation(eqn.slope, eqn.intercept + value);
         }
-        public static LinearEquation operator -(LinearEquation eqn, float value)
+        public static LinearEquation operator -(LinearEquation eqn, double value)
         {
             return new LinearEquation(eqn.slope, eqn.intercept - value);
         }
@@ -47,18 +47,18 @@ namespace Raytracer
 
     struct QuadraticEquation
     {
-        public float quadCoefficient;
-        public float linearCoefficient;
-        public float constant;
+        public double quadCoefficient;
+        public double linearCoefficient;
+        public double constant;
 
-        public QuadraticEquation(float quadCoefficient, float linearCoefficient, float constant)
+        public QuadraticEquation(double quadCoefficient, double linearCoefficient, double constant)
         {
             this.quadCoefficient = quadCoefficient;
             this.linearCoefficient = linearCoefficient;
             this.constant = constant;
         }
 
-        public float ValueAt(float t)
+        public double ValueAt(double t)
         {
             return (t * t * quadCoefficient) + (t * linearCoefficient) + constant;
         }
@@ -70,7 +70,7 @@ namespace Raytracer
                 eqn1.linearCoefficient + eqn2.linearCoefficient,
                 eqn1.constant + eqn2.constant);
         }
-        public static QuadraticEquation operator -(QuadraticEquation eqn, float value)
+        public static QuadraticEquation operator -(QuadraticEquation eqn, double value)
         {
             return new QuadraticEquation(
                 eqn.quadCoefficient,
